requirements.txt
django==2.1.1
django-heroku
gunicorn
django-debug-toolbar
elasticsearch-dsl>=6.0,<7.0
